moving ArgumentCollection to FakeItEasy.Configuration
