+++ Source/FakeItEasy-SL/FakeItEasy-SL.csproj
@@ -99,9 +99,6 @@
     <Compile Include="..\FakeItEasy\A.of.T.cs">
       <Link>A.of.T.cs</Link>
     </Compile>
-    <Compile Include="..\FakeItEasy\ArgumentCollection.cs">
-      <Link>ArgumentCollection.cs</Link>
-    </Compile>
     <Compile Include="..\FakeItEasy\ArgumentConstraintManagerExtensions.cs">
       <Link>ArgumentConstraintManagerExtensions.cs</Link>
     </Compile>
@@ -194,6 +191,9 @@
     </Compile>
     <Compile Include="..\FakeItEasy\Configuration\AnyCallConfiguration.cs">
       <Link>Configuration\AnyCallConfiguration.cs</Link>
+    </Compile>
+    <Compile Include="..\FakeItEasy\Configuration\ArgumentCollection.cs">
+      <Link>Configuration\ArgumentCollection.cs</Link>
     </Compile>
     <Compile Include="..\FakeItEasy\Configuration\BuildableCallRule.cs">
       <Link>Configuration\BuildableCallRule.cs</Link>
