+++ Source/FakeItEasy.Tests/Core/DefaultFakeObjectCallFormatterTests.cs
@@ -4,6 +4,7 @@ namespace FakeItEasy.Tests.Core
     using System.Diagnostics.CodeAnalysis;
     using System.Globalization;
     using System.Reflection;
+    using FakeItEasy.Configuration;
     using FakeItEasy.Core;
     using NUnit.Framework;
