+++ Source/FakeItEasy.Tests/Core/FakeCallEqualityComparerTests.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using System.Reflection;
+    using FakeItEasy.Configuration;
     using FakeItEasy.Core;
     using FakeItEasy.Tests.TestHelpers;
     using FluentAssertions;
