+++ Source/FakeItEasy.Tests/Expressions/ExpressionCallRuleTests.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using System.Linq.Expressions;
+    using FakeItEasy.Configuration;
     using FakeItEasy.Core;
     using FakeItEasy.Expressions;
     using FakeItEasy.Tests.TestHelpers;
