+++ Source/FakeItEasy.Tests/SelfInitializedFakes/RecordingManagerTests.cs
@@ -3,6 +3,7 @@
     using System.Collections.Generic;
     using System.Linq;
     using System.Reflection;
+    using FakeItEasy.Configuration;
     using FakeItEasy.Core;
     using FakeItEasy.SelfInitializedFakes;
     using FakeItEasy.Tests.TestHelpers;
