+++ Source/FakeItEasy/Configuration/ArgumentCollection.cs
@@ -1,4 +1,4 @@
-namespace FakeItEasy
+namespace FakeItEasy.Configuration
 {
     using System;
     using System.Collections;
